updated main error boundry... no more table flips
